Warnings

+++ modules/gin_simd/mipp/mipp_impl_SSE.hxx
@@ -2017,12 +2017,12 @@
 
     template <>
     inline msk cmpneq<int16_t>(const reg v1, const reg v2) {
-        return andnb<N<int16_t>()>(cmpeq<int16_t>(v1, v2), _mm_castps_si128(set1<int16_t>(0xFFFF)));
+        return andnb<N<int16_t>()>(cmpeq<int16_t>(v1, v2), _mm_castps_si128(set1<int16_t>(static_cast<int16_t>(0xFFFF))));
     }
 
     template <>
     inline msk cmpneq<int8_t>(const reg v1, const reg v2) {
-        return andnb<N<int8_t>()>(cmpeq<int8_t>(v1, v2), _mm_castps_si128(set1<int8_t>(0xFF)));
+        return andnb<N<int8_t>()>(cmpeq<int8_t>(v1, v2), _mm_castps_si128(set1<int8_t>(static_cast<int8_t>(0xFF))));
     }
 #endif
